Map conmtext typescript
