+++ package.json
@@ -11,6 +11,10 @@
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.5.0",
+    "@types/jest": "^27.0.2",
+    "@types/node": "^16.11.6",
+    "@types/react": "^17.0.33",
+    "@types/react-dom": "^17.0.10",
     "axios": "^0.21.4",
     "bootstrap": "^5.0.0-alpha3",
     "bootstrap-icons": "^1.4.0",
@@ -33,6 +37,7 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "swagger-typescript-api": "^6.4.0",
+    "typescript": "^4.4.4",
     "web-vitals": "^0.2.4",
     "yup": "^0.32.9"
   },
